almost there

+++ models/transaction.js
@@ -1,7 +1,7 @@
 module.exports = function(sequelize, DataTypes) {
   var Transactions = sequelize.define("Transactions", {
     Balance: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(15,2),
       allowNull: false,
       validate: {
         isNumeric: true,
@@ -9,7 +9,7 @@ module.exports = function(sequelize, DataTypes) {
       }
     },
     Amount: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(15,2),
       allowNull: false,
       validate: {
         isNumeric: true,
